backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1